refactor: finish moving all user-facing code to provider
